% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_pcv.R
\name{clean_pcv}
\alias{clean_pcv}
\title{Process SV File}
\usage{
clean_pcv(sv_file, project_code)
}
\arguments{
\item{sv_file}{Path to the SV file.}

\item{project_code}{Code for the project.}
}
\value{
A processed data frame containing the calculated values.
}
\description{
This function processes an SV file and performs various data manipulations and calculations.
}
