% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_water.R
\name{clean_water}
\alias{clean_water}
\title{Analyze Water Data}
\usage{
clean_water(water_data, project_code)
}
\arguments{
\item{water_data}{Water data, either a data frame or a CSV file.}

\item{project_code}{Project code for naming the output file.}
}
\value{
Processed data frame with water analysis results.
}
\description{
This function analyzes water data provided as a data frame or a CSV file.
}
