% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProjectFolders.R
\name{createProjectFolders}
\alias{createProjectFolders}
\title{Create project folders}
\usage{
createProjectFolders(projectDir)
}
\arguments{
\item{projectDir}{Character string specifying the project directory path.}
}
\value{
None
}
\description{
This function creates the project directories and subdirectories if they don't already exist.
}
\examples{
createProjectFolders("path/to/project")
}
