% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_h2_treatment.R
\name{get_h2_treatment}
\alias{get_h2_treatment}
\title{Calculate Heritability for a Treatment}
\usage{
get_h2_treatment(obj, pheno.col.no)
}
\arguments{
\item{obj}{The data frame containing the phenotype data.}

\item{pheno.col.no}{The column number of the phenotype of interest in the data frame.}
}
\value{
The heritability value for the treatment or NA if an error occurs.
}
\description{
This function calculates the heritability for a given treatment using linear mixed-effects modeling.
}
