% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash_1by1_vectorized.R
\name{mash_1by1_vectorized}
\alias{mash_1by1_vectorized}
\title{Perform MASH analysis for multiple columns simultaneously}
\usage{
mash_1by1_vectorized(data, alpha = 0, ...)
}
\arguments{
\item{data}{A list containing Bhat and Shat matrices.}

\item{alpha}{The tuning parameter controlling the amount of sparsity. Default is 0.}
}
\value{
A list containing posterior matrices and log likelihood.
}
\description{
This function performs MASH analysis using the ashr package for multiple columns
of Bhat and Shat matrices simultaneously. It calculates posterior mean,
posterior standard deviation, local false sign rate (lfsr), and log likelihood.
}
