% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_pcv_water.R
\name{merge_pcv_water}
\alias{merge_pcv_water}
\title{Merge pcv_data and water_data based on plantbarcode and day}
\usage{
merge_pcv_water(pcv_data, water_data, project_code)
}
\arguments{
\item{pcv_data}{Data frame containing pcv data}

\item{water_data}{Data frame containing water data}

\item{project_code}{Code for the project}
}
\value{
Merged data frame with additional calculated columns
}
\description{
Merge pcv_data and water_data based on plantbarcode and day
}
