% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performCluster.R
\name{performCluster}
\alias{performCluster}
\title{Perform clustering and extract data for each cluster}
\usage{
performCluster(hc_sbd, dat_pivoted_geno, joined_data, k = 10)
}
\arguments{
\item{hc_sbd}{Hierarchical clustering object.}

\item{dat_pivoted_geno}{Pivoted genotype data frame.}

\item{joined_data}{Joined data frame containing information about plants.}

\item{k}{Number of clusters to create.}
}
\value{
A list of data frames, where each data frame contains information about plants
belonging to a specific cluster.
}
\description{
This function takes a hierarchical clustering object, a pivoted genotype data frame,
and a joined data frame, and performs clustering to create k clusters. It then extracts
the corresponding data for each cluster and returns a list of data frames.
}
\seealso{
\code{\link{cutree}}, \code{\link{future_map}}, \code{\link{filter}}
}
