% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processTraits.R
\name{processTraits}
\alias{processTraits}
\title{Process Traits}
\usage{
processTraits(project_code, x)
}
\arguments{
\item{project_code}{Project code for naming the output files.}

\item{x}{Traits data, either a data frame or a CSV file.}
}
\value{
Processed data frames with traits analysis results.
}
\description{
This function processes traits data provided as a data frame or a CSV file.
}
