% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_process_csv.R
\name{read_and_process_csv}
\alias{read_and_process_csv}
\title{Read and Process CSV Files}
\usage{
read_and_process_csv(csv_dir, pattern)
}
\arguments{
\item{csv_dir}{Character string. The directory where the CSV files are located.}

\item{pattern}{Character string. The pattern to match the CSV files.}
}
\value{
A data frame containing the combined data from the CSV files.
}
\description{
This function reads and processes CSV files from a specified directory that
match a given pattern. It extracts information from the filenames and adds
it as new columns to the data frames. It then combines all the data frames
into a single data frame.
}
